refactor(sections): Move Banner-Section to sections folder and split by responsibility
