+++ src/components/sections/Banner/Description.tsx
@@ -1,0 +1,12 @@
+export function Description() {
+  return (
+    <div className="row flex flex-wrap w-full sm:justify-center sm:items-center lg:justify-normal lg:items-start mt-3 lg:mt-6 px-6 lg:px-0">
+      <div className="w-full md:w-3/4">
+        <p className="text-ebony text-lg">
+          I am a positive person who always seeks to compete with myself in
+          order to become the best version of myself.
+        </p>
+      </div>
+    </div>
+  );
+}
