+++ src/pages/HomePage/HomePage.tsx
@@ -1,9 +1,13 @@
-import HomePageBanner from "./Banner-Section";
+import HomePageBanner from "../../components/sections/Banner/HomePageBanner";
+import StackSection from "../../components/sections/Stack/StackSection";
 
 export default function HomePage() {
   return (
-    <div className="container mx-auto mt-5">
-      <HomePageBanner />
+    <div>
+      <div className="container mx-auto mt-5">
+        <HomePageBanner />
+      </div>
+      <StackSection />
     </div>
   );
 }
